Add files via upload
New-Style or C3PL for network access control on cisco switches running 16.x code.  We like to create a policy per vlan that we have on each switch so with the help of another script I have using ntc templates it will create a policy for all active vlans minus vlan 1 and make me a nice config to upload to the switch  Check out the ntmiko folder I have.