Add files via upload
We are standardizing a Voice VLAN at multiple sites and since it is all layer 2 we are keeping the same ID and name.  This helps when we authenticate the voip phones at these sites.

When it comes to authenticating these devices we hard code the voice vlan but we have a small caveat.  Some sites have a mix of 2 voip systems.  When we have Cisco phones at a site we hard code all ports for Cisco and let ISE figure out the what phone is what and dynamically assign the hosted voip vlan if necessary over the past year it has not been an issue and has saved us from having to reconfigure ports for the voice vlan.  That is why it is standardized across our sites.  Our result profile has the vlan name in it.